+++ src/components/Stepper/interfaces.ts
@@ -1,6 +1,7 @@
 export type StepperItem = {
   key: string;
   label: string;
+  onClick?: () => void;
 };
 
 export type StepperProps = {